requirements.txt
Flask==2.3.3
yt-dlp
Pillow
requests
loguru
python-dotenv
Flask-CORS